+++ src/main/java/lotto/service/LottoMachine.java
@@ -4,8 +4,6 @@ import java.util.ArrayList;
 import java.util.List;
 import lotto.domain.Lotto;
 import lotto.domain.PurchasedLottos;
-import lotto.domain.Rank;
-import lotto.domain.WinningLotto;
 import lotto.util.ErrorMessage;
 import lotto.util.NumberGenerate;
 
@@ -33,32 +31,6 @@ public class LottoMachine {
         
         this.money = money;
         return new PurchasedLottos(lottos);
-    }
-
-    public LottoResult calculateLottoWins(PurchasedLottos purchasedLotto, WinningLotto winningLotto) {
-        List<Rank> ranks = new ArrayList<>();
-        for (Lotto lotto : purchasedLotto.lottos()) {
-            boolean bonus = false;
-            List<Integer> lottoNums = lotto.numbers();
-
-            long count = sameNumberCountOf(winningLotto, lottoNums);
-            if (lottoNums.contains(winningLotto.bonusNumber())) {
-                bonus = true;
-            }
-            ranks.add(calculateRank(count, bonus));
-        }
-        return LottoResult.from(ranks);
-    }
-
-    private long sameNumberCountOf(WinningLotto winningLotto, List<Integer> lottoNums) {
-        return winningLotto.numbers()
-                .stream()
-                .filter(lottoNums::contains)
-                .count();
-    }
-
-    private Rank calculateRank(long count, boolean bonus) {
-        return Rank.calculateRank(count, bonus);
     }
 
     private int calculateLottoCount(int money) {
